requirements.txt
Django==4.2.15
requests==2.32.3
python-dotenv==1.0.1
python-telegram-bot==21.5